✨ feat(settings): add Developer settings group
Introduce a dedicated "Developer" group in the settings sidebar that
collects developer-facing surfaces: Advanced (first), API Key
(system-level, dev mode), OAuth Apps, and a new standalone Labs page.

- Extract the Labs section out of the Advanced page into its own
  `labs` settings tab; Advanced now only holds Tools & Diagnostics and
  App Updates.
- Move the system-level API Key, OAuth Apps and Advanced out of the
  System group into the always-visible Developer group.
- Wire the new Labs tab into the web + desktop component maps and the
  settings search index; add the `group.developer` i18n key.
