+++ locales/ar/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "اكتشف المزيد من الطرق لاستخدام LobeHub",
   "group.aiConfig": "النموذج",
   "group.common": "عام",
+  "group.developer": "مطور",
   "group.profile": "الحساب",
   "group.subscription": "الاشتراك",
   "group.system": "النظام",
