+++ locales/bg-BG/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Открийте повече начини за използване на LobeHub",
   "group.aiConfig": "Агент",
   "group.common": "Общи",
+  "group.developer": "Разработчик",
   "group.profile": "Акаунт",
   "group.subscription": "Абонамент",
   "group.system": "Система",
