+++ locales/de-DE/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Entdecken Sie weitere Möglichkeiten, LobeHub zu nutzen",
   "group.aiConfig": "Agent",
   "group.common": "Allgemein",
+  "group.developer": "Entwickler",
   "group.profile": "Konto",
   "group.subscription": "Abonnement",
   "group.system": "System",
