+++ locales/en-US/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Discover more ways to use LobeHub",
   "group.aiConfig": "Agent",
   "group.common": "General",
+  "group.developer": "Developer",
   "group.profile": "Account",
   "group.subscription": "Plans",
   "group.system": "System",
