+++ locales/es-ES/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Descubre más formas de usar LobeHub",
   "group.aiConfig": "Agente",
   "group.common": "General",
+  "group.developer": "Desarrollador",
   "group.profile": "Cuenta",
   "group.subscription": "Suscripción",
   "group.system": "Sistema",
