+++ locales/fa-IR/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "راه‌های بیشتری برای استفاده از LobeHub کشف کنید",
   "group.aiConfig": "عامل",
   "group.common": "عمومی",
+  "group.developer": "توسعه‌دهنده",
   "group.profile": "حساب کاربری",
   "group.subscription": "اشتراک",
   "group.system": "سیستم",
