+++ locales/fr-FR/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Découvrez plus de façons d'utiliser LobeHub",
   "group.aiConfig": "Agent",
   "group.common": "Général",
+  "group.developer": "Développeur",
   "group.profile": "Compte",
   "group.subscription": "Abonnement",
   "group.system": "Système",
