+++ locales/it-IT/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Scopri altri modi per utilizzare LobeHub",
   "group.aiConfig": "Agente",
   "group.common": "Generale",
+  "group.developer": "Sviluppatore",
   "group.profile": "Account",
   "group.subscription": "Abbonamento",
   "group.system": "Sistema",
