+++ locales/ja-JP/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "LobeHubを活用する方法をもっと発見",
   "group.aiConfig": "エージェント",
   "group.common": "共通",
+  "group.developer": "開発者",
   "group.profile": "アカウント",
   "group.subscription": "サブスクリプション",
   "group.system": "システム",
