+++ locales/ko-KR/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "LobeHub을 사용하는 더 많은 방법 발견하기",
   "group.aiConfig": "에이전트",
   "group.common": "일반",
+  "group.developer": "개발자",
   "group.profile": "계정",
   "group.subscription": "구독",
   "group.system": "시스템",
