+++ locales/nl-NL/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Ontdek meer manieren om LobeHub te gebruiken",
   "group.aiConfig": "Agent",
   "group.common": "Algemeen",
+  "group.developer": "Ontwikkelaar",
   "group.profile": "Account",
   "group.subscription": "Abonnement",
   "group.system": "Systeem",
