+++ locales/pl-PL/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Odkryj więcej sposobów korzystania z LobeHub",
   "group.aiConfig": "Agent",
   "group.common": "Ogólne",
+  "group.developer": "Programista",
   "group.profile": "Konto",
   "group.subscription": "Subskrypcja",
   "group.system": "System",
