+++ locales/pt-BR/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Descubra mais formas de usar o LobeHub",
   "group.aiConfig": "Agente",
   "group.common": "Geral",
+  "group.developer": "Desenvolvedor",
   "group.profile": "Conta",
   "group.subscription": "Assinatura",
   "group.system": "Sistema",
