+++ locales/ru-RU/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Откройте больше способов использовать LobeHub",
   "group.aiConfig": "Агент",
   "group.common": "Общие",
+  "group.developer": "Разработчик",
   "group.profile": "Аккаунт",
   "group.subscription": "Подписка",
   "group.system": "Система",
