+++ locales/tr-TR/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "LobeHub'u kullanmanın daha fazla yolunu keşfedin",
   "group.aiConfig": "Ajans",
   "group.common": "Genel",
+  "group.developer": "Geliştirici",
   "group.profile": "Hesap",
   "group.subscription": "Abonelik",
   "group.system": "Sistem",
