+++ locales/vi-VN/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "Khám phá thêm cách sử dụng LobeHub",
   "group.aiConfig": "Mô hình",
   "group.common": "Chung",
+  "group.developer": "Nhà phát triển",
   "group.profile": "Tài khoản",
   "group.subscription": "Gói đăng ký",
   "group.system": "Hệ thống",
