+++ locales/zh-CN/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "探索 LobeHub 的更多使用方式",
   "group.aiConfig": "智能体",
   "group.common": "通用",
+  "group.developer": "开发者",
   "group.profile": "账号",
   "group.subscription": "套餐",
   "group.system": "系统",
