+++ locales/zh-TW/setting.json
@@ -400,6 +400,7 @@
   "downloads.title": "探索更多使用 LobeHub 的方式",
   "group.aiConfig": "智能代理",
   "group.common": "通用",
+  "group.developer": "開發者",
   "group.profile": "帳號",
   "group.subscription": "訂閱",
   "group.system": "系統",
