+++ src/routes/(main)/settings/features/componentMap.desktop.ts
@@ -14,6 +14,7 @@ import Connector from '../connector';
 import Creds from '../creds';
 import Devices from '../devices';
 import Hotkey from '../hotkey';
+import Labs from '../labs';
 import Memory from '../memory';
 import Messenger from '../messenger';
 import OAuthApps from '../oauth-apps';
@@ -29,6 +30,7 @@ import SystemTools from '../system-tools';
 
 export const componentMap = {
   [SettingsTabs.Advanced]: Advanced,
+  [SettingsTabs.Labs]: Labs,
   [SettingsTabs.Appearance]: Appearance,
   [SettingsTabs.Provider]: Provider,
   [SettingsTabs.ServiceModel]: ServiceModel,
